dev-requirements.txt
dash[testing]>=2
requests[security]
flake8
flask
werkzeug
pytest
authlib
git+https://github.com/dhinesh03/flask-dance.git
